+++ client/src/calculator/Sub.java
@@ -8,21 +8,23 @@ import javax.xml.bind.annotation.XmlType;
 
 /**
  * <p>Java class for sub complex type.
- *
+ * 
  * <p>The following schema fragment specifies the expected content contained within this class.
- *
+ * 
  * <pre>
  * &lt;complexType name="sub">
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="arg0" type="{http://www.w3.org/2001/XMLSchema}int"/>
- *         &lt;element name="arg1" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *         &lt;element name="arg0" type="{http://www.w3.org/2001/XMLSchema}double"/>
+ *         &lt;element name="arg1" type="{http://www.w3.org/2001/XMLSchema}double"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
  *   &lt;/complexContent>
  * &lt;/complexType>
  * </pre>
+ *
+ *
  */
 @XmlAccessorType (XmlAccessType.FIELD)
 @XmlType (name = "sub", propOrder = {
@@ -31,34 +33,34 @@ import javax.xml.bind.annotation.XmlType;
 })
 public class Sub {
 
-	protected int arg0;
-	protected int arg1;
+	protected double arg0;
+	protected double arg1;
 
 	/**
 	 * Gets the value of the arg0 property.
 	 */
-	public int getArg0 () {
+	public double getArg0 () {
 		return arg0;
 	}
 
 	/**
 	 * Sets the value of the arg0 property.
 	 */
-	public void setArg0 (int value) {
+	public void setArg0 (double value) {
 		this.arg0 = value;
 	}
 
 	/**
 	 * Gets the value of the arg1 property.
 	 */
-	public int getArg1 () {
+	public double getArg1 () {
 		return arg1;
 	}
 
 	/**
 	 * Sets the value of the arg1 property.
 	 */
-	public void setArg1 (int value) {
+	public void setArg1 (double value) {
 		this.arg1 = value;
 	}
